Refactored temporary holder construction in converters.
The converters for class types would handle destruction of temporary
holder instances. This meant the destructor for every type being
converted needed to be unconditionally accessible.

This commit moves the destruction responsibility to
class_rep::convert_to(), which performs the holder construction anyway.

This makes it possible to use a dummy smart pointer to wrap
undestructible types, as can be seen in
test/test_private_destructors.cpp.