Add proper SONAME version and install/stage targets.
This adds the complete luabind version string to the shared library
names. It also adds a "d" suffix to the debug libraries. The build will
now produce names like:

 libluabind.so.0.8.1
 libluabindd.so.0.8.1

"bjam install" will build and install the libraries on the default
prefix (/usr/local on *nix). The --prefix option can be used to change
the prefix. The install also creates the unversioned symlinks:

 libluabind.so -> libluabind.so.0.8.1
 libluabindd.so -> libluabindd.so.0.8.1