Fix test_class_info.cpp.
The order in which lua_next returns the entries of a Lua table is unspecified,
but test/test_class_info.cpp only checked for the y attribute as first and x
attribute as second entry of the info.attributes table (which, in turn is
filled in by luabind::get_class_info by iterating (internally via lua_next)
over the objects class_rep table.
This lead to the test randomly failing with:
    test\test_class_info.cpp:49"ERROR: "[string "x = X()..."]:6: assertion failed!""
at the line "assert(info.attributes[1] == 'y')". Fixed to check for x and y
both as first and second members (as was already done with class_names()'s
return value).