build: Prefer source tree over installed headers.
Fix having to do "make install" before changes in headers would be applied to
the build, leading to inconsistent library/header configurations before a second
"make && make install". This situation arises when luabind and Boost are
installed to the same location, which they are by default on Linux
(/usr/local/include).