Partially revert ecfe25e for performance reasons.
ecfe25e: "Fix #25: Overload candidates overflow" was somewhat detrimental for
performance (~5x in MSVC Debug mode), since this code is run on every function
call. So instead of using a std::vector, revert to the statically allocated
10-element array but check for overflows. Since the array (except for the first
element if it is the only one) is only used to generate the ambiguity error
message, just discard addidional elements, keeping track of how many there were
and display that number in the error message if > 0.